backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guitar-transcriber"
version = "0.1.0"
description = "Guitar audio to tab transcription service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "basic-pitch>=0.3.0",
    "pretty-midi>=0.2.10",
    "pyguitarpro>=0.9",
    "numpy>=1.24.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
